feat(timeInterval): add higher-order lettable version of timeInterval
