+++ src/operators/index.ts
@@ -67,6 +67,7 @@ export { takeLast } from './takeLast';
 export { tap } from './tap';
 export { throttle } from './throttle';
 export { throttleTime } from './throttleTime';
+export { timeInterval } from './timeInterval';
 export { timestamp } from './timestamp';
 export { toArray } from './toArray';
 export { window } from './window';
